243: move FeeHistoryParams and FeeHistory
