+++ src/rpc_client/requests.rs
@@ -1,4 +1,4 @@
-use crate::rpc_client::eth_rpc::{BlockSpec, FixedSizeData};
+use crate::rpc_client::eth_rpc::{BlockSpec, FixedSizeData, Quantity};
 use ic_ethereum_types::Address;
 use serde::Serialize;
 use candid::Deserialize;
@@ -35,4 +35,31 @@ pub struct GetLogsParam {
     /// Each topic can also be an array of DATA with "or" options.
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub topics: Vec<Vec<FixedSizeData>>,
+}
+
+/// Parameters of the [`eth_feeHistory`](https://ethereum.github.io/execution-apis/api-documentation/) call.
+#[derive(Debug, Serialize, Clone)]
+#[serde(into = "(Quantity, BlockSpec, Vec<u8>)")]
+pub struct FeeHistoryParams {
+    /// Number of blocks in the requested range.
+    /// Typically providers request this to be between 1 and 1024.
+    pub block_count: Quantity,
+    /// Highest block of the requested range.
+    /// Integer block number, or "latest" for the last mined block or "pending", "earliest" for not yet mined transactions.
+    pub highest_block: BlockSpec,
+    /// A monotonically increasing list of percentile values between 0 and 100.
+    /// For each block in the requested range, the transactions will be sorted in ascending order
+    /// by effective tip per gas and the corresponding effective tip for the percentile
+    /// will be determined, accounting for gas consumed.
+    pub reward_percentiles: Vec<u8>,
+}
+
+impl From<FeeHistoryParams> for (Quantity, BlockSpec, Vec<u8>) {
+    fn from(value: FeeHistoryParams) -> Self {
+        (
+            value.block_count,
+            value.highest_block,
+            value.reward_percentiles,
+        )
+    }
 }
