+++ src/rpc_client/responses.rs
@@ -200,3 +200,30 @@ impl HttpResponsePayload for Block {
         Some(ResponseTransform::Block)
     }
 }
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct FeeHistory {
+    /// Lowest number block of the returned range.
+    #[serde(rename = "oldestBlock")]
+    pub oldest_block: BlockNumber,
+    /// An array of block base fees per gas.
+    /// This includes the next block after the newest of the returned range,
+    /// because this value can be derived from the newest block.
+    /// Zeroes are returned for pre-EIP-1559 blocks.
+    #[serde(rename = "baseFeePerGas")]
+    pub base_fee_per_gas: Vec<WeiPerGas>,
+    /// An array of block gas used ratios (gasUsed / gasLimit).
+    #[serde(default)]
+    #[serde(rename = "gasUsedRatio")]
+    pub gas_used_ratio: Vec<f64>,
+    /// A two-dimensional array of effective priority fees per gas at the requested block percentiles.
+    #[serde(default)]
+    #[serde(rename = "reward")]
+    pub reward: Vec<Vec<WeiPerGas>>,
+}
+
+impl HttpResponsePayload for FeeHistory {
+    fn response_transform() -> Option<ResponseTransform> {
+        Some(ResponseTransform::FeeHistory)
+    }
+}
